Add EpiModel base class
