+++ episimlab/models/__init__.py
@@ -1,0 +1,1 @@
+from .epi_model import EpiModel
