--- episimlab/models/toy/__init__.py
@@ -1,1 +1,0 @@
-from .markov import markov
